refactor: split the cart page into separate components
